Experimenting with nodevtools and manual build in Spack
nodevtools seems to work well to bootstap a recent version of
gcc and compiles flux. However, same issue with flux run
not working (and flux broker downstream, too).

What if I can trick Spack into using the dependencies I've
installed for the builing of flux WITHOUT using the flux
package itself?  I want to build Flux with the libraries in
Spack. It seems to sort of work - the hang up is detecting
the location of include (.h) header files for the libraries
that will be linked to Flux.