Flux-in-Spack install uses Flux's Python, not miniconda3 Python
When Flux is installed with Spack, a special version of Python is
installed. Here, we activate that Python and install Parsl into the
Python that is needed for Flux, thus bypassing Conda and minimizing
the number of potentially conflicting Python installations. On the
other hand, this also forces us to implicitly load Parsl.